requirements.txt
asgiref==3.7.2
Django==4.2.3
django-appconf==1.0.5
django-autoslug==1.9.9
django-cleanup==8.0.0
django-environ==0.10.0
django-imagekit==4.1.0
django-utils==0.0.2
pilkit==2.0
Pillow==10.0.0
psycopg2==2.9.9
psycopg2-binary==2.9.7
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
whitenoise==6.7.0
